backend/requirements.txt
# === Core Framework ===
fastapi
uvicorn[standard]
pydantic

# === Database ===
sqlalchemy
psycopg2-binary
asyncpg

# === Data Processing ===
pandas
pyarrow
numpy
yfinance

# === Object Storage ===
minio

# === Message Queue ===
pika

# === Caching ===
redis

# === HTTP / Networking ===
aiohttp
requests
websockets

# === Web Scraping & NLP ===
beautifulsoup4
vaderSentiment

# === AI / ML ===
google-generativeai
huggingface_hub

# === Monitoring ===
prometheus-fastapi-instrumentator
slowapi

# === Auth ===
bcrypt
PyJWT
passlib[bcrypt]
python-multipart
python-jose[cryptography]

# === Email ===
fastapi-mail==1.4.1

# === Workflow ===
inngest

# === Broker API (Shoonya) ===
NorenRestApiPy
websocket-client
pyotp

# === Utilities ===
python-dotenv
apscheduler
greenlet
feedparser
